+++ src/application/structs/Options.h
@@ -2,8 +2,10 @@
 
 enum Goal
 {
+    GoalLevels,
     GoalEmblems,
     GoalEmeraldHunt,
+    GoalLevelsAndEmeraldHunt,
     GoalEmblemsAndEmeraldHunt,
     None
 };