+++ src/output/locationRepository/LocationRepository.h
@@ -1,6 +1,9 @@
 ﻿#pragma once
 #include <map>
 #include "../../application/structs/LocationData.h"
+#include "../../application/structs/Options.h"
+#include "../../application/structs/LevelStatus.h"
+
 
 
 class LocationRepository
@@ -11,14 +14,10 @@ public:
     LocationData GetLocation(int checkId);
     std::map<int, LocationData> GetLocations();
     std::vector<LifeBoxLocationData> GetLifeCapsules();
+    bool CompletedAllLevels(Options options);
+    LevelStatus GetLevelStatus(Options options);
 
 private:
-    std::string GetLocationName(Characters character, LevelIDs levelId, int mission);
-    LocationData GetLocationFromUpgrade(EventFlags eventFlags, const char* str, int i);
-    LocationData GetLocationFromLevel(Characters character, LevelIDs levelId, int mission);
-    LocationData GetLocationFromSubLevel(EventFlags eventFlags, const char* str, int i);
-    LocationData GetLocationFromFieldEmblem(EventFlags eventFlags, const char* str, int i);
-    LocationData GetLocationFromEvent(EventFlags eventFlags, const char* str, int i);
     std::map<int, LocationData> _checkData;
     std::vector<LifeBoxLocationData> _lifeCapsules;
 };